@@ -1,8 +1,7 @@
 {
-  "name": "simplesceneeditor",
-  "version": "0.1.0",
+  "name": "editor",
+  "version": "0.0.0",
   "description": "A simple 3D scene editor based on WebGL",
-  "private": true,
   "repository": {
     "type": "git",
     "url": "git+https://github.com/KuTuGu/SimpleSceneEditor.git"
@@ -19,42 +18,32 @@
   },
   "homepage": "https://github.com/KuTuGu/SimpleSceneEditor#readme",
   "scripts": {
-    "start": "npm run serve",
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "test:unit": "vue-cli-service test:unit"
-  },
-  "dependencies": {
-    "core-js": "^2.6.5",
-    "vue": "^2.6.10",
-    "vue-router": "^3.1.3",
-    "vuex": "^3.1.1"
-  },
-  "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.11.0",
-    "@vue/cli-plugin-eslint": "^3.11.0",
-    "@vue/cli-plugin-unit-jest": "^4.1.2",
-    "@vue/cli-service": "^4.1.2",
-    "@vue/eslint-config-prettier": "^5.0.0",
-    "@vue/test-utils": "1.0.0-beta.29",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^25.0.0",
-    "eslint": "^5.16.0",
-    "eslint-plugin-prettier": "^3.1.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "lint-staged": "^8.1.5",
-    "prettier": "^1.18.2",
-    "vue-template-compiler": "^2.6.10"
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint --fix --ext .js,.ts,.vue src/"
   },
   "gitHooks": {
     "pre-commit": "lint-staged"
   },
   "lint-staged": {
-    "*.{js,vue}": [
-      "vue-cli-service lint",
-      "git add"
+    "src/**/*.{js,ts,vue}": [
+      "npm run lint"
     ]
+  },
+  "dependencies": {
+    "vue": "^3.0.2",
+    "vue-router": "^4.0.0-rc.2",
+    "vuex": "^4.0.0-rc.1"
+  },
+  "devDependencies": {
+    "@vue/compiler-sfc": "^3.0.2",
+    "eslint": "^7.12.1",
+    "eslint-config-prettier": "^6.15.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-vue": "^7.1.0",
+    "lint-staged": "^10.5.1",
+    "prettier": "^2.1.2",
+    "vite": "^1.0.0-rc.8",
+    "yorkie": "^2.0.0"
   }
 }
