@@ -167,5 +167,5 @@ export {
   TEXTURE_VSHADER_SOURCE,
   TEXTURE_FSHADER_SOURCE,
   COLOR_VSHADER_SOURCE,
-  COLOR_FSHADER_SOURCE
+  COLOR_FSHADER_SOURCE,
 };
